+++ CHANGELOG.md
@@ -3,4 +3,4 @@
 #### v1.0.0 - Dec 26, 2023
 
 - added IntervalCounter with tests
-- added IntervalTimer (wip)
+- added IntervalTimer with tests